pyproject.toml
[tool.poetry]
name = "graphrah-query"
version = "0.1.0"
description = ""
authors = ["6ixGODD <6goddddddd@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    { include = "graphrag_query", from = "." }
]


[tool.poetry.dependencies]
python = "^3.11"
tiktoken = "^0.8.0"
typing-extensions = "^4.12.2"
openai = "^1.51.2"
pandas = "1.5"
numpy = "1.24"
pydantic-settings = "^2.5.2"
json-repair = "^0.30.0"
pyarrow = "^17.0.0"
lancedb = "^0.14.0"
loguru = { version = "^0.7.2", optional = true }
pyqt6 = { version = "^6.7.1", optional = true }
fastapi = { version = "^0.115.0", optional = true }
markdown = { version = "^3.7", optional = true }
tabulate = {version = "^0.9.0", optional = true}
uvicorn = {version = "^0.31.1", optional = true}
gunicorn = {version = "^23.0.0", optional = true}


[tool.poetry.group.logging.dependencies]
loguru = "^0.7.2"

[tool.poetry.group.gui.dependencies]
pyqt6 = "^6.7.1"
markdown = "^3.7"

[tool.poetry.group.web.dependencies]
fastapi = "^0.115.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
